+++ src/main/java/com/example/helios/member/repository/HospitalRepository.java
@@ -1,6 +1,6 @@
 package com.example.helios.member.repository;
 
-import com.example.helios.member.domain.Hospital;
+import com.example.helios.member.entity.Hospital;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,7 +10,7 @@ import java.util.Optional;
 
 //테스트용
 @Repository
-public interface HospitalRepository extends JpaRepository<Hospital, Integer> {
+public interface HospitalRepository extends JpaRepository<Hospital, Long> {
 
     // 이메일로 병원 계정 조회 (로그인 시 사용)
     Optional<Hospital> findByEmail(String email);
